teams-deploy/requirements.txt
botbuilder-core>=4.14.0
botbuilder-integration-aiohttp>=4.14.0
azure-ai-projects==2.0.0b3
azure-identity>=1.15.0
openai>=1.0.0
azure-storage-blob>=12.19.0